requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
email-validator==2.1.0.post1
Flask-Login==0.6.3
WTForms==3.1.2
Werkzeug==3.0.1
bootstrap-flask==2.3.3